requirements.txt
streamlit
pandas
python-dotenv
langchain
langchain-community
langchain-groq
sentence-transformers
faiss-cpu
PyMuPDF